+++ src/libs/__tests__/pr.test.ts
@@ -5,7 +5,7 @@ import { handlePullRequestMessage } from '../pr';
 const comments = [
   {
     id: 2,
-    body: '#### :tropical_drink: `preview` on myFirstProject/staging. <summary>Pulumi report</summary>',
+    body: '#### :tropical_drink: `preview` on myFirstProject/staging. <summary>Pulumi report</summary><!-- pulumi-comment-id: some-comment-staging-id -->',
   },
 ];
 const resp = { data: comments };
@@ -50,7 +50,7 @@ describe('pr.ts', () => {
 
     await handlePullRequestMessage(defaultOptions, projectName, 'test');
     expect(createComment).toHaveBeenCalledWith({
-      body: '#### :tropical_drink: `preview` on myFirstProject/staging\n\n<details>\n<summary>Pulumi report</summary>\n\n```\ntest\n```\n\n</details>',
+      body: '#### :tropical_drink: `preview` on myFirstProject/staging\n\n<details>\n<summary>Pulumi report</summary>\n\n```\ntest\n```\n\n</details>\n<!-- pulumi-comment-id: default -->',
       issue_number: 123,
     });
   });
@@ -142,7 +142,31 @@ describe('pr.ts', () => {
     await handlePullRequestMessage(options, projectName, 'test');
     expect(updateComment).toHaveBeenCalledWith({
       comment_id: 2,
-      body: '#### :tropical_drink: `preview` on myFirstProject/staging\n\n<details>\n<summary>Pulumi report</summary>\n\n```\ntest\n```\n\n</details>',
+      body: '#### :tropical_drink: `preview` on myFirstProject/staging\n\n<details>\n<summary>Pulumi report</summary>\n\n```\ntest\n```\n\n</details>\n<!-- pulumi-comment-id: default -->',
+    });
+  });
+
+  it('should edit the comment if it finds a previous created one by the comment id', async () => {
+    // @ts-ignore
+    gh.context = {
+      payload: {
+        pull_request: {
+          number: 123,
+        },
+      },
+    };
+
+    const options = {
+      command: 'preview',
+      stackName: 'staging',
+      commentOnPrId: 'some-comment-staging-id',
+      editCommentOnPr: true,
+    } as Config;
+
+    await handlePullRequestMessage(options, projectName, 'test');
+    expect(updateComment).toHaveBeenCalledWith({
+      comment_id: 2,
+      body: '#### :tropical_drink: `preview` on myFirstProject/staging\n\n<details>\n<summary>Pulumi report</summary>\n\n```\ntest\n```\n\n</details>\n<!-- pulumi-comment-id: some-comment-staging-id -->',
     });
   });
 });